% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_calculate_cost_distance.R
\name{calculate_cost_distance}
\alias{calculate_cost_distance}
\title{Calculate least-cost paths between points}
\usage{
calculate_cost_distance(raster, points, x_col, y_col)
}
\arguments{
\item{raster}{A SpatRaster representing resistance values}

\item{points}{A data frame containing point coordinates}

\item{x_col}{Name of the longitude column in points}

\item{y_col}{Name of the latitude column in points}
}
\value{
An sf object with least-cost paths and cost distances
}
\description{
This function computes least-cost paths between all points based on a
resistance raster using the leastcostpath package.
}
\examples{
library(terra)
library(sf)
library(leastcostpath)
r <- rast(nrows = 20, ncols = 20)
values(r) <- runif(ncell(r)) + 0.1
points <- data.frame(
  x = runif(4, xmin(r), xmax(r)),
  y = runif(4, ymin(r), ymax(r))
)
lcps <- calculate_cost_distance(r, points, "x", "y")
plot(r)
plot(sf::st_geometry(lcps), add = TRUE, col = "red", lwd = 2)
}
