% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lcp.R
\name{calculate_lcp}
\alias{calculate_lcp}
\title{Calculate least-cost path between two points}
\usage{
calculate_lcp(raster, from, to)
}
\arguments{
\item{raster}{A SpatRaster representing resistance values}

\item{from}{A numeric vector with coordinates of the start point (x, y)}

\item{to}{A numeric vector with coordinates of the end point (x, y)}
}
\value{
An sf object representing the least-cost path
}
\description{
This function computes the least-cost path between two locations based on a
resistance raster using the leastcostpath package.
}
\examples{
library(terra)
library(sf)
library(leastcostpath)
r <- rast(nrows = 20, ncols = 20)
values(r) <- runif(ncell(r)) + 0.1
from <- c(xmin(r), ymin(r))
to   <- c(xmax(r), ymax(r))
lcp <- calculate_lcp(r, from, to)
plot(r, col = terrain.colors(50), main = "Least Cost Path")
points(sf::st_coordinates(sf::st_geometry(lcp)), col = "red", pch = 16)
plot(sf::st_geometry(lcp), add = TRUE, col = "blue", lwd = 3)
}
