% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_extract_env.R
\name{extract_env}
\alias{extract_env}
\title{Extract environmental values for point data}
\usage{
extract_env(points, raster, x_col, y_col)
}
\arguments{
\item{points}{A data frame containing point coordinates}

\item{raster}{A SpatRaster object with environmental variables}

\item{x_col}{Name of the longitude column in points}

\item{y_col}{Name of the latitude column in points}
}
\value{
A data frame combining the original points with extracted raster values
}
\description{
This function extracts environmental raster values at given point locations.
}
\examples{
library(terra)
r1 <- rast(nrows=10, ncols=10)
values(r1) <- runif(ncell(r1))
r2 <- rast(nrows=10, ncols=10)
values(r2) <- runif(ncell(r2))
env <- c(r1, r2)
points <- data.frame(
  x = runif(5, xmin(env), xmax(env)),
  y = runif(5, ymin(env), ymax(env))
)
extract_env(points, env, "x", "y")
}
