% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sdm.R
\name{predict_sdm}
\alias{predict_sdm}
\title{Predict species distribution on raster data}
\usage{
predict_sdm(model, raster, type = "response")
}
\arguments{
\item{model}{A fitted model object (e.g. from fit_sdm_glm)}

\item{raster}{A SpatRaster object with environmental predictor variables}

\item{type}{Type of prediction (default = "response")}
}
\value{
A SpatRaster with predicted suitability values
}
\description{
This function applies a fitted SDM model to environmental raster data
and returns a raster of predicted suitability values.
}
\examples{
library(terra)
r1 <- rast(nrows=20, ncols=20)
values(r1) <- runif(ncell(r1))
r2 <- rast(nrows=20, ncols=20)
values(r2) <- runif(ncell(r2))
env <- c(r1, r2)
names(env) <- c("temp", "precip")
points <- data.frame(
  x = runif(50, xmin(env), xmax(env)),
  y = runif(50, ymin(env), ymax(env)),
  presence = rbinom(50, 1, 0.5)
)
data_model <- extract_env(points, env, "x", "y")
model <- fit_sdm_glm(
  data = data_model,
  response = "presence",
  predictors = c("temp", "precip")
)
pred_raster <- predict_sdm(model, env)
plot(pred_raster)
}
