% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudo_absence.R
\name{generate_pseudo_absence}
\alias{generate_pseudo_absence}
\title{Generate pseudo-absence points}
\usage{
generate_pseudo_absence(data, x_col, y_col, n, min_dist = NULL)
}
\arguments{
\item{data}{A data frame containing presence points with coordinates}

\item{x_col}{Name of the longitude column}

\item{y_col}{Name of the latitude column}

\item{n}{Number of pseudo-absence points to generate}

\item{min_dist}{Minimum distance to presence points (optional)}
}
\value{
A data frame with presence (1) and pseudo-absence (0)
}
\description{
This function creates pseudo-absence points based on presence data.
}
